% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibGP.R
\name{calibGP}
\alias{calibGP}
\title{A function to process calibration using a completed DAE.
This is an S4 class to represent TROLL surrogated calibration.}
\usage{
calibGP(
  yobs,
  xobs,
  xcalib,
  paramsFormulas,
  covFormulas,
  surmodel,
  ncores,
  verbose = TRUE,
  calib.opts = NULL
)
}
\arguments{
\item{yobs}{d.f. Observed summary statitics.}

\item{xobs}{d.f. Observed variables for available summary statitics.}

\item{xcalib}{d.f. Available variable on the whole studied setup.}

\item{paramsFormulas}{list. list of equations linking estimated parameters.}

\item{covFormulas}{d.f. Available covariable for 'paramsFormulas'
on the whole studied setup.}

\item{surmodel}{list. an adjusted TROLL surrogate model.}

\item{ncores}{int. number of core to use.}

\item{verbose}{bool. allow verbose}

\item{calib.opts}{list. Calibrations options.}
}
\description{
A function to process calibration using a completed DAE.
This is an S4 class to represent TROLL surrogated calibration.
}
