% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeGP.R
\name{computeGP}
\alias{computeGP}
\title{Gaussian process modeling with heteroskedastic noise}
\usage{
computeGP(
  dae,
  lower = NULL,
  upper = NULL,
  covtype = NULL,
  noiseControl = NULL,
  settings = NULL,
  inmem = FALSE
)
}
\arguments{
\item{dae}{trolldae. an initialized trolldae object.}

\item{lower, upper}{num. optional bounds for the theta parameter
(see \code{hetGP::\link[hetGP:mleHetGP]{mleHetGP}} for details).}

\item{covtype}{covariance kernel type, either 'Gaussian', 'Matern5_2' or 'Matern3_2'.
(see \code{hetGP::\link[hetGP:mleHetGP]{mleHetGP}} for details)}

\item{noiseControl}{list with elements related to optimization of the noise process parameters.
(see \code{hetGP::\link[hetGP:mleHetGP]{mleHetGP}} for details)}

\item{settings}{list for options about the general modeling procedure
(see \code{hetGP::\link[hetGP:mleHetGP]{mleHetGP}} for details)}

\item{inmem}{bool. Save dae within surmodel object.}
}
\value{
a surmodel object.
}
\description{
Gaussian process modeling with heteroskedastic noise
}
