% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeHM.R
\name{computeHM}
\alias{computeHM}
\title{History matching approach with hierarchical
emulators (\code{\link[hmer:variance_emulator_from_data]{hmer}}) for stochastic systems}
\usage{
computeHM(dae, targets, genPts = 100, ratioVal = 0.1, inmem = FALSE)
}
\arguments{
\item{dae}{trolldae. an initialized trolldae object.}

\item{targets}{list. list of target boundaries.
A pair (min, max), where min represents the lower bound and max
the upper bound for the target.
(see \code{hetGP::\link[hetGP:mleHetGP]{mleHetGP}} for details)}

\item{genPts}{int. number of points to be generated within
plausible space}

\item{ratioVal}{num. percentage of validation points for emulator.}

\item{inmem}{bool. Save dae within surmodel object.}
}
\value{
a surmodel object.
}
\description{
History matching approach with hierarchical
emulators (\code{\link[hmer:variance_emulator_from_data]{hmer}}) for stochastic systems
}
