% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{corcorr}
\alias{corcorr}
\title{internal Correlation Matrix Correction}
\usage{
corcorr(vars, cor, l, FLAGSTOP)
}
\arguments{
\item{vars}{mat. WIP LHS matrix.}

\item{cor}{mat. Correlation matrix.}

\item{l}{int. Convergence tolerance.}

\item{FLAGSTOP}{int. Flag to check convergence .}
}
\value{
Void with outputs files written in the defined folder.
}
\description{
This function changes the order in which data is organized in order
to force the correlation matrix to a prescribed value. This implementation
uses the Hungtington-Lyrintzis algorithm.
}
