% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lhscorcorr.R
\name{.lhscorcorr}
\alias{.lhscorcorr}
\title{Correlation matrix Correction}
\usage{
.lhscorcorr(vars, cormat = 0, eps = 0.005, echo = FALSE, maxit = 0)
}
\arguments{
\item{vars}{df. mat. "raw" Latin Hypercube Sample.}

\item{cormat}{mat. A Pearson correlation matrix.}

\item{eps}{num. Tolerance deviation.}

\item{echo}{bool. Verbose.}

\item{maxit}{int. maximum number of iterations.
(0: use a heuristic based on the size of the hypercube, -1: no stop)}
}
\value{
df. the correlation matrix corrected.
}
\description{
Corrects the correlation matrix of a given Latin Hypercube Sample.
}
\details{
This function changes the order in which data is organized in order
to force the correlation matrix to a prescribed value. This implementation
uses the Hungtington-Lyrintzis algorithm.
}
\examples{
require(SLHD)
uncorlhs <- maximinSLHD(t = 1,m = 100,k = 2)$StandDesign
corm <- matrix(c(1,0.5,0.5,1),2,2)
.lhscorcorr(vars = uncorlhs,cormat = corm)
.lhscorcorr(vars = uncorlhs,cormat = corm,maxit = 1E10)

}
\references{
Chalom, A. and Prado, P.I.K.L. 2012. Parameter space exploration of ecological models
\emph{arXiv}:1210.6278
}
