% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilsStack.R
\name{gatherStack}
\alias{gatherStack}
\alias{gatherStack,list-method}
\title{Function to gather trollstack into a list of trollsim}
\usage{
gatherStack(listSim, name, ...)

\S4method{gatherStack}{list}(listSim, name, ...)
}
\arguments{
\item{listSim}{trollstack.}

\item{name}{chr. name of the stack.}

\item{...}{Unused parameters.}
}
\value{
a list of trollsim
}
\description{
Function to gather trollstack into a list of trollsim
}
