% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processExp.R
\name{processExperiment}
\alias{processExperiment}
\title{Process an individual experiment on a TROLL simulation.}
\usage{
processExperiment(sim, singleExp, parameters = NULL, inputs = NULL)
}
\arguments{
\item{sim}{trollsim. a TROLL simulaiton object to process.}

\item{singleExp}{trollexpsingle. a trollexpsingle object to be applied on 'sim'.}

\item{parameters}{list.list of trollexpsingle object.}

\item{inputs}{list. Initial inputs for experiments.}
}
\value{
a trollexpsingle object
}
\description{
Process an individual experiment on a TROLL simulation.
}
\examples{

suppressWarnings(suppressMessages(library(dplyr)))

fnExpSum <- function(x,...){
   summary <- as.matrix(data.frame("DBH" = mean(x@forest$dbh),
   "LAI" = mean(x@forest$LAI)),nrow = 1,byrow = TRUE)
   return(summary)
   }
   
fnExpFrac <- function(x,parameters,...){
   x@forest <-  x@forest \%>\% sample_frac(parameters$fraction) 
   return(x)
   }  
   
Exp1 <- createExperiment(id = 1, 
 type = "Inter",
 deltaT = 100,
 fnExp = fnExpFrac,
 parameters = data.frame("fraction" = 0.5),
 inputs = list())
  
   
Exp2 <- createExperiment(id = 2, 
 type = "Summary", 
 fnExp = fnExpSum,
 parameters = data.frame(),
 inputs = list())
 
 simTest <- TROLLv3_sim
 
 processExperiment(sim = simTest,
 singleExp = Exp1,parameters = data.frame("fraction" = 0.5))
 
 processExperiment(sim = simTest,
 singleExp = Exp2)

}
