% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilsStack.R
\name{splitStack}
\alias{splitStack}
\alias{splitStack,trollstack-method}
\title{Function to split trollstack into a list of trollsim}
\usage{
splitStack(stack, ...)

\S4method{splitStack}{trollstack}(stack, ...)
}
\arguments{
\item{stack}{list. a trollstack to split.}

\item{...}{Unused parameters.}
}
\value{
a list of trollsim
}
\description{
Function to split trollstack into a list of trollsim
}
