% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surcalib.R
\docType{class}
\name{surcalib-class}
\alias{surcalib-class}
\title{An S4 class to represent TROLL surrogated calibration.
This is an S4 class to represent TROLL surrogated calibration.}
\description{
An S4 class to represent TROLL surrogated calibration.
This is an S4 class to represent TROLL surrogated calibration.
}
\section{Slots}{

\describe{
\item{\code{yobs}}{d.f. Observed summary statitics.}

\item{\code{xobs}}{d.f. Observed variables for available summary statitics.}

\item{\code{xcalib}}{d.f. Available variable on the whole studied setup.}

\item{\code{ycalib}}{d.f. Estimated summary statitics on the whole studied setup.}

\item{\code{paramCalib}}{d.f Estimated parameters values on the whole studied setup.}

\item{\code{paramsFormulas}}{list. list of equations linking estimated parameters.}

\item{\code{covFormulas}}{d.f. Available covariable for 'paramCalib'
on the whole studied setup.}

\item{\code{surmodel}}{list. an adjusted TROLL surrogate model.}

\item{\code{calib.opts}}{list. Calibrations options.}
}}

