% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surcalib.R
\name{surcalib}
\alias{surcalib}
\title{An S4 class to represent TROLL surrogated calibration.
This is an S4 class to represent TROLL surrogated calibration.}
\usage{
surcalib(
  yobs = data.frame(),
  xobs = data.frame(),
  xcalib = data.frame(),
  params.formulas = list(),
  cov.formulas = data.frame(),
  surmodel = list(),
  calib.opts = list()
)
}
\arguments{
\item{yobs}{d.f. Observed summary statitics.}

\item{xobs}{d.f. Observed variables for available summary statitics.}

\item{xcalib}{d.f. Available variable on the whole studied setup.}

\item{params.formulas}{list. list of equations linking estimated parameters.}

\item{cov.formulas}{d.f. Available covariable for 'params.formulas'
on the whole studied setup.}

\item{surmodel}{list. an adjusted TROLL surrogate model.}

\item{calib.opts}{list. Calibrations options.}
}
\description{
An S4 class to represent TROLL surrogated calibration.
This is an S4 class to represent TROLL surrogated calibration.
}
