% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surmodel.R
\docType{class}
\name{surmodel-class}
\alias{surmodel-class}
\title{An S4 class to represent TROLL surrogate model.
This is an S4 class to represent TROLL surrogate model.}
\description{
An S4 class to represent TROLL surrogate model.
This is an S4 class to represent TROLL surrogate model.
}
\section{Slots}{

\describe{
\item{\code{type}}{char. Type of selected surrgogate model.
(None/ hetGP)}

\item{\code{lhs}}{matrix. Latin Hypercube sample used for the design.}

\item{\code{xsim}}{matrix. unscaled tested values.}

\item{\code{ysim}}{matrix. unscaled response.}

\item{\code{params}}{df. Tested parameters with design specifications.}

\item{\code{models}}{list. list of adjusted models.}

\item{\code{models.opts}}{list. list of model specifications.}

\item{\code{dae}}{list. summary statitics on design and experiments.}

\item{\code{check}}{bool. Post-hoc validation of surrogate models.}
}}

