% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surmodel.R
\name{surmodel}
\alias{surmodel}
\title{An S4 class to represent TROLL surrogate model.
This is an S4 class to represent TROLL surrogate model.}
\usage{
surmodel(
  type = character(),
  lhs = matrix(),
  xsim = matrix(),
  ysim = matrix(),
  params = data.frame(),
  models = list(),
  modelsopts = list(),
  dae = list(),
  check = FALSE
)
}
\arguments{
\item{type}{char. Type of selected surrgogate model.
(None/ hetGP)}

\item{lhs}{matrix. Latin Hypercube sample used for the design.}

\item{xsim}{matrix. unscaled tested values.}

\item{ysim}{matrix. unscaled response.}

\item{params}{df. Tested parameters with design specifications.}

\item{models}{list. list of adjusted models.}

\item{modelsopts}{list. list of model specifications.}

\item{dae}{list. summary statitics on design and experiments.}

\item{check}{bool. Post-hoc validation of surrogate models.}
}
\description{
An S4 class to represent TROLL surrogate model.
This is an S4 class to represent TROLL surrogate model.
}
