% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trolldae.R
\docType{class}
\name{trolldae-class}
\alias{trolldae-class}
\title{An S4 class to represent TROLL design and simulations.
This is an S4 class to represent TROLL design and the simulations.}
\description{
An S4 class to represent TROLL design and simulations.
This is an S4 class to represent TROLL design and the simulations.
}
\details{
'boundaries' data.frame is formatted as followed:
\itemize{
\item parameter: chr. Name of the parameter within TROLL settings,
covariate table or experiment setup ;
\item quantileFn: fct. Quantile function to scale margin
distribution of studied parameter ;
\item type: chr. Type of parameter ('global'/ 'species'/
'climate'/ 'daily'/ 'covariate/ 'experiment').
}
}
\section{Slots}{

\describe{
\item{\code{name}}{char. Design setup name.}

\item{\code{state}}{char. Current state of design
('Generated'/ 'Initialized'/ 'Pre-simulation'/ 'In-processing'/ 'Post-simulation').}

\item{\code{lhs}}{matrix. Latin Hypercube sample used for the design.}

\item{\code{boundaries}}{d.f. a formatted data.frame (see details) to
describe parameters space.}

\item{\code{xsim}}{matrix. unscaled tested values.}

\item{\code{ysim}}{matrix. unscaled response (if summary experiment provided).}

\item{\code{params}}{df. Tested parameters with design specifications.}

\item{\code{doeopts}}{list. Options for initial simulation.}

\item{\code{experiments}}{list. Setup of experiments to apply.}

\item{\code{path}}{char. Path to the simulation(s).}

\item{\code{type}}{char. Type of design
(Raw : without model, GP: Gaussian processes assisted).}

\item{\code{simopts}}{list. List of  simulations options.}
}}

