% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trolldae.R
\name{trolldae}
\alias{trolldae}
\title{An S4 class to represent TROLL design and the simulations.}
\usage{
trolldae(
  name = character(),
  state = NA_character_,
  lhs = matrix(),
  boundaries = data.frame(),
  xsim = matrix(),
  ysim = matrix(),
  params = data.frame(),
  doeopts = list(ntotalsim = NA_integer_, nsim = NA_integer_, nreplica = NA_integer_,
    corrmat = matrix(), fnParams = list(global = NULL, climate = NULL, daily = NULL,
    lidar = NULL), forestInit = data.frame(), dimPlot = list()),
  experiments = list(),
  path = character(),
  type = character(),
  simopts = list(nyearsInit = NA_integer_, nyearsSampling = NA_integer_, nyearsStep =
    NA_integer_, nyearsMax = NA_integer_)
)
}
\arguments{
\item{name}{char. Design setup name.}

\item{state}{char. char. Current state of design
('Generated'/ 'Initialized'/ 'Pre-simulation'/ 'In-processing'/ 'Post-simulation').}

\item{lhs}{matrix. Latin Hypercube sample used for the design.}

\item{boundaries}{d.f. a formatted data.frame (see details) to
describe parameters space.}

\item{xsim}{matrix. unscaled tested values.}

\item{ysim}{matrix. unscaled response (if summary experiment provided).}

\item{params}{df. Tested params with design specifications.}

\item{doeopts}{list. Options for initial simulation.}

\item{experiments}{list. Setup of experiments to apply.}

\item{path}{char. Path to the simulation(s).}

\item{type}{char. Type of design
(Raw : without model, GP: Gaussian processes assisted).}

\item{simopts}{list. List of  simulations options.}
}
\description{
This is an S4 class to represent TROLL design and the simulations.
}
\details{
'boundaries' data.frame is formatted as followed:
\itemize{
\item parameter: chr. Name of the parameter within TROLL settings,
covariate table or experiment setup ;
\item quantileFn: fct. Quantile function to scale margin
distribution of studied parameter ;
\item type: chr. Type of parameter ('global'/ 'species'/
'climate'/ 'daily'/ 'covariate/ 'experiment').
}
}
