% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trolldesign.R
\docType{class}
\name{trolldesign-class}
\alias{trolldesign-class}
\title{An S4 class to represent TROLL design and simulations.
This is an S4 class to represent TROLL design and the simulations.}
\description{
An S4 class to represent TROLL design and simulations.
This is an S4 class to represent TROLL design and the simulations.
}
\section{Slots}{

\describe{
\item{\code{name}}{char. Design setup name.}

\item{\code{lhs}}{matrix. Latin Hypercube sample used for the design.}

\item{\code{parameters}}{df. Tested parameters with design specifications.}

\item{\code{initopts}}{list. Options for initial simulation.}

\item{\code{experiments}}{list. list of experiments to apply.}

\item{\code{path}}{char. Path to the simulation(s).}

\item{\code{type}}{char. Type of design
(Raw : without model, GP: Gaussian processes).}

\item{\code{simopts}}{list. List of  simulations options.}
}}

