% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trolldesign.R
\name{trolldesign}
\alias{trolldesign}
\title{An S4 class to represent TROLL design and the simulations.}
\usage{
trolldesign(
  name = character(),
  lhs = matrix(),
  parameters = data.frame(),
  initopts = list(),
  experiments = list(),
  path = character(),
  type = character(),
  simopts = list()
)
}
\arguments{
\item{name}{char. Design setup name.}

\item{lhs}{matrix. Latin Hypercube sample used for the design.}

\item{parameters}{df. Tested parameters with design specifications.}

\item{initopts}{list. Options for initial simulation.}

\item{experiments}{list. list of experiments to apply.}

\item{path}{char. Path to the simulation(s).}

\item{type}{char. Type of design
(Raw : without model, GP: Gaussian processes).}

\item{simopts}{list. List of  simulations options.}
}
\description{
This is an S4 class to represent TROLL design and the simulations.
}
