% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trollexp.R
\docType{class}
\name{trollexp-class}
\alias{trollexp-class}
\title{An S4 class to represent TROLL experiments.}
\description{
This is an S4 class to represent TROLL experiments.
}
\section{Slots}{

\describe{
\item{\code{id}}{int. Experiment identifier.}

\item{\code{type}}{char. Type of experiment
("initialization"/ "intermediate"/ "summarize").}

\item{\code{func}}{func. experiment function to apply.}

\item{\code{parameters}}{char. list of used parameters.}

\item{\code{inputs}}{list. List of inputs objects.}

\item{\code{outputs}}{list. List of outputs objects.}
}}

