% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trollexp.R
\name{trollexp}
\alias{trollexp}
\title{An S4 class to represent TROLL design and the experiments.}
\usage{
trollexp(
  id = integer(),
  type = character(),
  func = function(x) {
     x
 },
  parameters = character(),
  inputs = list(),
  outputs = list()
)
}
\arguments{
\item{id}{int. Experiment identifier.}

\item{type}{char. Type of experiment.
("initialization"/ "intermediate"/ "summarize").}

\item{func}{func. experiment function to appy on trollsim/ trollstack}

\item{parameters}{char. list of used parameters}

\item{inputs}{list. List of inputs objects used in experiments}

\item{outputs}{list. List of outputs objects used in other simulations}
}
\description{
This is an S4 class to represent TROLL design and the experiments.
}
