% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trollsetup.R
\docType{class}
\name{trollexpsetup-class}
\alias{trollexpsetup-class}
\title{An S4 class to represent spedific setup of TROLL experiments.}
\description{
This is an S4 class to represent a setup of TROLL experiments.
}
\section{Slots}{

\describe{
\item{\code{params}}{d.f. vector of used parameters.}

\item{\code{listexp}}{list. list of TROLL experiments.}

\item{\code{deltat}}{int. cumulated simulation time.}

\item{\code{inputs.opts}}{list. List of inputs objects.}

\item{\code{outputs.opts}}{list. List of outputs objects.}
}}

