% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trollsetup.R
\name{trollexpsetup}
\alias{trollexpsetup}
\title{An S4 class to represent spedific setup of TROLL experiments.}
\usage{
trollexpsetup(
  params = data.frame(),
  listexp = list(),
  deltat = integer(),
  inputs.opts = list(),
  outputs.opts = list()
)
}
\arguments{
\item{params}{d.f. vector of used parameters.}

\item{listexp}{list. list of TROLL experiments.}

\item{deltat}{int. cumulated simulation time.}

\item{inputs.opts}{list. List of inputs objects.}

\item{outputs.opts}{list. List of outputs objects.}
}
\description{
This is an S4 class to represent a setup of TROLL experiments.
}
