% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trollexp.R
\docType{class}
\name{trollexpsingle-class}
\alias{trollexpsingle-class}
\title{An S4 class to represent TROLL experiments.}
\description{
This is an S4 class to represent a single TROLL experiment.
}
\section{Slots}{

\describe{
\item{\code{id}}{int. Experiment identifier.}

\item{\code{type}}{char. Type of experiment
("initialization"/ "intermediate"/ "summarize").}

\item{\code{func}}{func. experiment function to apply.}

\item{\code{deltat}}{int. Number of year to process after experiment.}

\item{\code{parameters}}{d.f. vector of used parameters.}

\item{\code{inputs.opts}}{list. List of inputs objects.}

\item{\code{outputs.opts}}{list. List of outputs objects.}
}}

