% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trollexp.R
\name{trollexpsingle}
\alias{trollexpsingle}
\title{An S4 class to represent TROLL design and the experiments.}
\usage{
trollexpsingle(
  id = integer(),
  type = character(),
  func = function(x) {
     x
 },
  deltat = integer(),
  parameters = data.frame(),
  inputs.opts = list(),
  outputs.opts = list()
)
}
\arguments{
\item{id}{int. Experiment identifier.}

\item{type}{char. Type of experiment
("initialization"/ "intermediate"/ "summarize").}

\item{func}{func. experiment function to apply.}

\item{deltat}{int. Number of year to process after experiment.}

\item{parameters}{d.f. vector of used parameters.}

\item{inputs.opts}{list. List of inputs objects.}

\item{outputs.opts}{list. List of outputs objects.s}
}
\description{
This is an S4 class to represent TROLL design and the experiments.
}
