% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TROLLv3_sim.R
\docType{data}
\name{TROLLv3_sim}
\alias{TROLLv3_sim}
\title{TROLL sim}
\format{
A trollsim object.
}
\usage{
TROLLv3_sim
}
\description{
A TROLL outputs from a 1-year simulation on a 100x100 grid with other parameters
are sets to default parameters and using TROLLv3_species, TROLLv3_climatedaytime12, and
TROLLv3_daytimevar for use in tests and examples. Ecosystem level output has
been thinned and species output has been removed to save disk space.
}
\keyword{datasets}
