% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TROLLv3_stack.R
\docType{data}
\name{TROLLv3_stack}
\alias{TROLLv3_stack}
\title{TROLL stack}
\format{
A trollstack object.
}
\usage{
TROLLv3_stack
}
\description{
A TROLL outputs from a 1-year simulation on a 100x100 grid with 50 000 and 500 annual constant seed rain.
Other parameters are sets todefault parameters and using TROLLv3_species, TROLLv3_climatedaytime12, and
TROLLv3_daytimevar for use in tests and examples. Ecosystem level output has
been thinned and species output has been removed to save disk space.
}
\keyword{datasets}
