% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibGP.R
\name{calibGP}
\alias{calibGP}
\title{An S4 class to represent TROLL surrogated calibration.
This is an S4 class to represent TROLL surrogated calibration.}
\usage{
calibGP(
  yobs,
  xobs,
  xcalib,
  params.formulas,
  cov.formulas,
  surmodel,
  calib.opts = NULL
)
}
\arguments{
\item{yobs}{d.f. Observed summary statitics.}

\item{xobs}{d.f. Observed variables for available summary statitics.}

\item{xcalib}{d.f. Available variable on the whole studied setup.}

\item{params.formulas}{list. list of equations linking estimated parameters.}

\item{cov.formulas}{d.f. Available covariable for 'params.formulas'
on the whole studied setup.}

\item{surmodel}{list. an adjusted TROLL surrogate model.}

\item{calib.opts}{list. Calibrations options.}
}
\description{
An S4 class to represent TROLL surrogated calibration.
This is an S4 class to represent TROLL surrogated calibration.
}
