% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trollcalibr-package.R
\docType{package}
\name{trollcalibr-package}
\alias{trollcalibr}
\alias{trollcalibr-package}
\title{trollcalibr: What the Package Does (One Line, Title Case)}
\description{
Function and code to calibrate the individual-based and spatially-explicit forest growth model TROLL.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/gsalzet/trollcalibr}
  \item \url{https://gsalzet.github.io/trollcalibr/}
  \item Report bugs at \url{https://github.com/gsalzet/trollcalibr/issues}
}

}
\author{
\strong{Maintainer}: Guillaume Salzet \email{guillaume.salzet@inrae.fr} (\href{https://orcid.org/0000-0003-4548-5673}{ORCID})

Authors:
\itemize{
  \item Sylvain Schmitt \email{sylvain.m.schmitt@gmail.com} (\href{https://orcid.org/0000-0001-7759-7106}{ORCID})
}

}
\keyword{internal}
